// src/modules/rental-vehicle/enums/rental-vehicle.enum.ts
export enum RentalVehicleApprovalStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  REJECTED = 'rejected',
  INACTIVE = 'inactive',
}

export enum RentalVehicleAvailabilityStatus {
  AVAILABLE = 'available',
  UNAVAILABLE = 'unavailable',
  MAINTENANCE = 'maintenance',
}

export enum RentalVehicleType {
  BIKE = 'bike',
  CAR = 'car',
}